+++ simulation/g4simulation/g4cemc/RawTowerCalibration.h
@@ -35,7 +35,7 @@ public:
     detector = d;
   }
   void
-  TowerType(const double type)
+  TowerType(const int type)
   {
     _tower_type = type; 
   } 
@@ -149,7 +149,7 @@ protected:
   double _zero_suppression_GeV;
 
   //! tower type to act on
-  double _tower_type; 
+  int _tower_type; 
 
   PHTimeServer::timer _timer;
